requirements.txt
streamlit==1.38
langchain-community
langchain-core
langchain-openai
pydantic==1.10.12
psycopg2-binary
database
python-dotenv
Werkzeug
